Crate p3_frontend

Conversion from a circuit that implements the Air trait into a halo2_backend compatible circuit. Includes helper functions to convert plonky3 witness format into halo2_backend witness format.

Structs

CompileParams

FWrap

PreprocessingInfo

Public

Query

SymbolicAirBuilder

An AirBuilder for evaluating constraints symbolically, and recording them for later use.

SymbolicVariable

A variable within the evaluation window, i.e. a column in either the local or next row.

Enums

Location

SymbolicExpression

An expression over SymbolicVariables.

Var

Traits

AirBuilderWithPublicValues

Functions

check_witness

compile_circuit_cs

compile_preprocessing

get_public_inputs

trace_to_wit